sub/requirements.txt
google-cloud-pubsub
google-cloud-logging
requests